hc_3900/requirements-min.txt
fastapi
uvicorn
polars
pydantic
sqlalchemy
sqlmodel
pyyaml
numpy
pandas
scikit-learn
joblib
umap-learn
matplotlib
